+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ## New features
 * `mwModule()` now return `controller` value, with possibility to use new `clear()` method
+* add `header` and `footer` arguments to `mwModuleUI()`
 
 # manipulateWidget 0.8.0 (2017-10-25)